+++ templates/mobile/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @vitdoc/template-mobile
 
+## 2.1.0-beta.11
+
+### Patch Changes
+
+- fix vitdoc ui alias
+
+## 2.1.0-beta.10
+
+### Patch Changes
+
+- upgrade
+
 ## 2.1.0-beta.9
 
 ### Patch Changes